second refactoring
