+++ lib/core/router/app_router.dart
@@ -19,7 +19,7 @@ import '../../features/calendar/data/models/calendar_model.dart';
 import '../../features/calendar/presentation/viewmodels/personal_calendar_view_model.dart';
 import '../../features/calendar/presentation/viewmodels/shared_calendar_view_model.dart';
 import '../../features/calendar/presentation/views/setting/calendar_setting_screen.dart';
-import '../../features/calendar/presentation/views/shared/ list/shared_calendar_list_screen.dart';
+import '../../features/calendar/presentation/views/shared/list/shared_calendar_list_screen.dart';
 import 'app_shell.dart';
 
 GoRouter createRouter(BuildContext context) {
