+++ lib/features/calendar/presentation/widgets/schedule_dialog/schedule_preview_card.dart
